telegram-scoring/requirements.txt
gspread==6.*
google-auth==2.*
openai==1.*
python-dotenv==1.*
pyyaml==6.*
jinja2==3.*
tenacity==9.*
python-docx==1.*
